separating api and pagination test concerns
