--- README.rdoc
@@ -1,31 +1,0 @@
-== README
-
-{<img src="https://travis-ci.org/pseudocms/app-api.svg" alt="Build
-Status" />}[https://travis-ci.org/pseudocms/app-api]
-
-This README would normally document whatever steps are necessary to get the
-application up and running.
-
-Things you may want to cover:
-
-* Ruby version
-
-* System dependencies
-
-* Configuration
-
-* Database creation
-
-* Database initialization
-
-* How to run the test suite
-
-* Services (job queues, cache servers, search engines, etc.)
-
-* Deployment instructions
-
-* ...
-
-
-Please feel free to use a different markup language if you do not plan to run
-<tt>rake doc:app</tt>.
