+++ test/support/api_test.rb
@@ -8,84 +8,15 @@ module APITest
         version
       end
 
-      send(:include, APITestHelpers)
-    end
-
-    def pagination_test(setup_method)
-      define_method :pagination_setup do |page, per_page|
-        send(setup_method, page, per_page)
-      end
-
-      send(:include, PagingTestHelpers)
+      send(:include, APITest::TestHelpers)
+      send(:include, APITest::TestModule)
     end
   end
 
-  module PagingTestHelpers
-    def test_pagination_for_first_page
-      pagination_setup(1, 1)
-      assert_response :ok
-
-      refute page_links.has_key?(:first)
-      refute page_links.has_key?(:prev)
-      assert page_links.has_key?(:next)
-      assert page_links.has_key?(:last)
-    end
-
-    def test_pagination_for_middle_page
-      pagination_setup(2, 1)
-      assert_response :ok
-
-      assert page_links.has_key?(:first)
-      assert page_links.has_key?(:prev)
-      assert page_links.has_key?(:next)
-      assert page_links.has_key?(:last)
-    end
-
-    def test_pagination_for_last_page
-      last_page = last_page_number
-      pagination_setup(last_page, 1)
-      assert_response :ok
-
-      assert page_links.has_key?(:first)
-      assert page_links.has_key?(:prev)
-      refute page_links.has_key?(:next)
-      refute page_links.has_key?(:last)
-    end
-
-    private
-
-    LINK_HEADER_PATTERN = /\A<([^>]+)>;\s*rel="(.*)"\z/
-
-    def page_links
-      @page_links ||= begin
-        headers = response.headers['Link'].split(',').map(&:strip)
-        {}.tap do |links|
-          headers.each do |header|
-            if LINK_HEADER_PATTERN =~ header
-              links[$2] = $1
-            end
-          end
-
-          links.symbolize_keys!
-        end
-      end
-    end
-
-    def last_page_number
-      last_page = 0
-
-      ActiveRecord::Base.transaction do
-        pagination_setup(1, 1)
-        last_page = page_links[:last][/page=(\d+)/][$1]
-        raise ActiveRecord::Rollback, "rolling it back"
-      end
-
-      @page_links = nil
-      last_page
-    end
+  module TestModule
   end
 
-  module APITestHelpers
+  module TestHelpers
 
     def post(uri, params = {}, headers = {})
       super(uri, params, default_headers.merge(headers))
@@ -141,8 +72,4 @@ module APITest
       @json_response ||= JSON.parse(response.body)
     end
   end
-end
-
-class ActionDispatch::IntegrationTest
-  include APITest
 end
