+++ test/support/core_ext/action_dispatch/integration_test.rb
@@ -1,0 +1,7 @@
+require "support/api_test"
+require "support/pagination_test"
+
+class ActionDispatch::IntegrationTest
+  include APITest
+  include PaginationTest
+end
