+++ test/support/core_ext/active_support/test_case.rb
@@ -1,0 +1,4 @@
+class ActiveSupport::TestCase
+  self.use_transactional_fixtures = true
+  include FactoryGirl::Syntax::Methods
+end
